requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=2.111=mkl
blas-devel=3.9.0=11_win64_mkl
ca-certificates=2021.9.30=haa95532_1
certifi=2020.12.5=py36haa95532_0
cudatoolkit=11.1.1=heb2d755_7
cycler=0.10.0=py36haa95532_0
dataclasses=0.8=pyh787bdff_2
freetype=2.10.4=h546665d_1
icu=58.2=ha925a31_3
intel-openmp=2021.3.0=h57928b3_3372
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py36hd77b12b_0
libblas=3.9.0=11_win64_mkl
libcblas=3.9.0=11_win64_mkl
liblapack=3.9.0=11_win64_mkl
liblapacke=3.9.0=11_win64_mkl
libpng=1.6.37=h1d00b33_2
libtiff=4.2.0=hd0e1b90_0
libuv=1.42.0=h8ffe710_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.3.4=py36haa95532_0
matplotlib-base=3.3.4=py36h49ac443_0
mkl=2021.3.0=hb70f87d_564
mkl-devel=2021.3.0=h57928b3_565
mkl-include=2021.3.0=hb70f87d_564
msys2-conda-epoch=20160418=1
ninja=1.10.2=h2d74725_1
numpy=1.17.5=py36hc71023c_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1l=h2bbff1b_0
pillow=8.3.1=py36h4fa10fc_0
pip=21.0.1=py36haa95532_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
python=3.6.13=h3758d61_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.9.1=py3.6_cuda11.1_cudnn8_0
qt=5.9.7=vc14h73c81de_0
setuptools=58.0.4=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
tbb=2021.3.0=h2d74725_0
tk=8.6.11=h8ffe710_1
torchaudio=0.9.1=py36
torchvision=0.10.1=py36_cu111
tornado=6.1=py36h2bbff1b_0
typing_extensions=3.10.0.2=pyha770c72_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_1
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
